Initial refactoring
* Created a new project to hold all the business logic, models etc
* Added seed data so we have some data to work with
