+++ Backend/TodoList.Api/Core/TodoList.Infrastructure/Data/TodoContext.cs
@@ -1,0 +1,22 @@
+using Microsoft.EntityFrameworkCore;
+using TodoList.Infrastructure.Data.Models;
+
+namespace TodoList.Infrastructure.Data
+{
+    public class TodoContext : DbContext
+    {
+        public TodoContext(DbContextOptions<TodoContext> options)
+            : base(options)
+        {
+        }
+
+        public DbSet<TodoItem> TodoItems => Set<TodoItem>();
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<TodoItem>().HasData(Seed.Todos);
+            
+            base.OnModelCreating(modelBuilder);
+        }
+    }
+}
