--- Backend/TodoList.Api/TodoList.Api/TodoItem.cs
@@ -1,13 +1,0 @@
-using System;
-
-namespace TodoList.Api
-{
-    public class TodoItem
-    {
-        public Guid Id { get; set; }
-
-        public string Description { get; set; }
-
-        public bool IsCompleted { get; set; }
-    }
-}
